Add option '--fmt' to provide C formatting options
Allow for a user provided format to be applied when generating the
sets with options --16 or --32.

By default, the format is preserved, but for example, allows
generating the same output for option --32 that was originally
implemented in a4f295a (Implement the --32 option, 2014-04-16)
when calling:

  $ uniset --32 --fmt=0x%04X

The user string is restricted to a small size and validated to
prevent abuse, including not allowing writes and keeping reads
within 32 bits (which should be ok for systems that have that
width for ints), but is ultimately the user's responsability to
get the generator right, so all rope will be provided for their
hanging, as requested.