autotools: error in configure if no C99 compiler found
While the check for its availability is already present, if none is
found, then whatever compiler was detected will be used, resulting in a
broken build.

The codebase indeed requires C99 and even a GNU extension or C11 to
build, so make sure to exit with an error in that case.